+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [1.50.1] - 2021-11-17
+
 ### Fixed
 
 - Strip ':<port>' suffix when requesting a client certificate.
@@ -630,7 +632,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.50.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.50.1...HEAD
+[1.50.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.50.0...v1.50.1
 [1.50.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.49.0...v1.50.0
 [1.49.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.48.1...v1.49.0
 [1.48.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.48.0...v1.48.1